Add files via upload
The presentations are attached for details.
Limited comments (due to time constraints), hence will be making a video maybe soon explaining the whole program.
References are given in the code. 
This code is mine and is utterly modified and is made into a "wholly new game".
For honesty: No offense to anyone, this was a minor group project in which more than 85% to 90% of coding/testing in it as well as the presentations were done/made by me.
The code has some anomalies which I took care of but not absolutely. The anomalies' basis lies in the ball-bar as well as the in the ball-brick collisions. Will try to make a video explaining all of them soon whenever time permits me. 
If you want an explanation beforehand, please write to me at: hrithikchugh007@gmail.com